requirements.txt
absl-py==0.7.1
aioredis==1.3.0
amqp==2.5.2
appnope==0.1.0
asgi-redis==1.4.3
asgiref==3.2.3
asn1crypto==1.1.0
astor==0.8.0
async-timeout==2.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.13.0
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-background-task==0.1.8
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==3.2.0
django-heroku==0.3.1
django-redis==4.10.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==20.0.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.2.0
incremental==17.5.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonfield==2.0.2
jsonschema==3.0.1
kiwisolver==1.1.0
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.0
msgpack==0.6.2
msgpack-python==0.5.6
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
python-decouple==3.3
python-hglib==2.6.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
whitenoise==4.1.4
widgetsnbextension==3.5.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
zipp==0.6.0
zope.interface==4.6.0
